# backend/openapi-schema.yml
openapi: 3.0.3
info:
  title: Ks Financial App API
  version: 1.0.0
  description: |2

    ## API для системы учёта и анализа персональных финансов

    ### Аутентификация
    - Используйте **POST /api/v1/auth/login/** для получения токенов
    - Access token передавайте в заголовке: `Authorization: Bearer <token>`
    - Refresh token автоматически устанавливается в HttpOnly cookie

    ### Основные эндпоинты
    - **Транзакции** - CRUD операций с финансами
    - **Категории** - управление категориями
    - **Аналитика** - статистика и AI рекомендации
    - **Блог** - посты и комментарии

    ### AI Функционал
    - **AI Рекомендации** - персональные советы по экономии
    - **SMS Парсинг** - автоматическое создание транзакций из SMS
paths:
...
  /api/v1/auth/refresh/:
    post:
      operationId: v1_auth_refresh_create
      description: |-
        Обновление токена из Cookie.
        POST /api/v1/auth/refresh/
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefreshRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
...
components:
  schemas:
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
    TokenRefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...